feat: created HallucinationDetector and HallucinationEvaluator classes
